requirements.txt
requests
google-cloud-bigquery
google-cloud-firestore
google-cloud-storage